[#272] refactor: path 이름 변경
